requirement.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
scikit-learn==1.3.2
joblib==1.3.2
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.1
python-multipart==0.0.6
